pyproject.toml
[project]
name = "telegram-f1-top-points-mcp-bot"
version = "0.1.0"
description = "Telegram bot for F1 standings with email delivery via FastMCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.0.2",
    "python-telegram-bot>=22.1",
    "fastmcp>=0.2.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
]
